+++ rfc_parameters.h
@@ -35,25 +35,26 @@ rfc_set_value_return_t rfc_set_parameter_value(RFC_FUNCTION_HANDLE function_hand
                                zval *value);
 
 
-zval rfc_get_char_value(DATA_CONTAINER_HANDLE h, SAP_UC *name, unsigned len);
+zval rfc_get_char_value(DATA_CONTAINER_HANDLE h, SAP_UC *name, unsigned len, unsigned char rtrim_enabled);
 zval rfc_get_date_value(DATA_CONTAINER_HANDLE h, SAP_UC *name);
 zval rfc_get_time_value(DATA_CONTAINER_HANDLE h, SAP_UC *name);
 zval rfc_get_byte_value(DATA_CONTAINER_HANDLE h, SAP_UC *name, unsigned len);
-zval fc_get_table_value(DATA_CONTAINER_HANDLE h, SAP_UC *name);
+zval fc_get_table_value(DATA_CONTAINER_HANDLE h, SAP_UC *name, unsigned char rtrim_enabled);
 zval rfc_get_num_value(DATA_CONTAINER_HANDLE h, SAP_UC *name, unsigned len);
 zval rfc_get_float_value(DATA_CONTAINER_HANDLE h, SAP_UC *name);
 zval rfc_get_int_value(DATA_CONTAINER_HANDLE h, SAP_UC *name);
 zval rfc_get_int1_value(DATA_CONTAINER_HANDLE h, SAP_UC *name);
 zval rfc_get_int2_value(DATA_CONTAINER_HANDLE h, SAP_UC *name);
-zval rfc_get_structure_value(DATA_CONTAINER_HANDLE h, SAP_UC *name);
+zval rfc_get_structure_value(DATA_CONTAINER_HANDLE h, SAP_UC *name, unsigned char rtrim_enabled);
 zval rfc_get_string_value(DATA_CONTAINER_HANDLE h, SAP_UC *name);
 zval rfc_get_xstring_value(DATA_CONTAINER_HANDLE h, SAP_UC *name);
 
-zval rfc_get_table_line(RFC_STRUCTURE_HANDLE line, zend_string *param_name);
-zval rfc_get_field_value(RFC_STRUCTURE_HANDLE h, RFC_FIELD_DESC field_desc);
+zval rfc_get_table_line(RFC_STRUCTURE_HANDLE line, zend_string *param_name, unsigned char rtrim_enabled);
+zval rfc_get_field_value(RFC_STRUCTURE_HANDLE h, RFC_FIELD_DESC field_desc, unsigned char rtrim_enabled);
 
 zval rfc_get_parameter_value(RFC_FUNCTION_HANDLE function_handle,
                                RFC_FUNCTION_DESC_HANDLE function_desc_handle,
-                               zend_string *name);
+                               zend_string *name,
+                               unsigned char rtrim_enabled);
 
 #endif /* _RFC_PARAMETERS_H */